Create Module-UpdateStudentFormRows
This module was created to abstract the function of setting a summary row on a form (summary row displays info in form controls like current rank, payment status and displaying a progress bar). Summary rows existed on 2 pages: the Students page and the Student Details page. The summary row has now been removed from the Student Details page due to a visual revamp, but this function is still large enough to have its own module and may be reused in the future.