Create Class-ColorCollection
Created this module for easy reuse in other modules. I was using these colors repeatedly in different modules and I always had to make a comment to know what color was being applied, then search down the color when I needed it again. Colors can now be applied as shown in the code comments.